refactor(sdk): extract GraphQLExecutor protocol and HttpxGraphQLExecutor adapter
Introduce a narrow owned GraphQLExecutor Protocol and rename BasePipefyClient to
HttpxGraphQLExecutor, the sole httpx/gql adapter that implements it. base_client
stays as a transitional alias so services keep inheriting until each is migrated
to constructor injection.

Move unwrap_relay_connection_nodes to utils/relay and _format_internal_api_error
to services/internal_api_errors so neither depends on the executor module. Add a
mock_executor test double alongside mock_internal_api_client. No behavior change.
