+++ packages/sdk/tests/services/test_pipefy_base_client.py
@@ -32,8 +32,8 @@ async def test_execute_query_builds_transport_with_tls_verification(valid_settin
     mock_session.execute = AsyncMock(return_value={"ok": True})
 
     with (
-        patch("pipefy_sdk.base_client.HTTPXAsyncTransport") as mock_transport_cls,
-        patch("pipefy_sdk.base_client.Client") as mock_client_cls,
+        patch("pipefy_sdk.graphql_executor.HTTPXAsyncTransport") as mock_transport_cls,
+        patch("pipefy_sdk.graphql_executor.Client") as mock_client_cls,
     ):
         mock_client_cls.return_value.__aenter__ = AsyncMock(return_value=mock_session)
         mock_client_cls.return_value.__aexit__ = AsyncMock(return_value=None)
@@ -55,7 +55,7 @@ async def test_execute_query_passes_variables_to_session(valid_settings):
     mock_session = AsyncMock()
     mock_session.execute = AsyncMock(return_value={"ok": True})
 
-    with patch("pipefy_sdk.base_client.Client") as mock_client_cls:
+    with patch("pipefy_sdk.graphql_executor.Client") as mock_client_cls:
         mock_client_cls.return_value.__aenter__ = AsyncMock(return_value=mock_session)
         mock_client_cls.return_value.__aexit__ = AsyncMock(return_value=None)
 
@@ -81,7 +81,7 @@ async def test_execute_query_omits_variable_values_for_empty_dict(valid_settings
     mock_session = AsyncMock()
     mock_session.execute = AsyncMock(return_value={"ok": True})
 
-    with patch("pipefy_sdk.base_client.Client") as mock_client_cls:
+    with patch("pipefy_sdk.graphql_executor.Client") as mock_client_cls:
         mock_client_cls.return_value.__aenter__ = AsyncMock(return_value=mock_session)
         mock_client_cls.return_value.__aexit__ = AsyncMock(return_value=None)
 
@@ -114,7 +114,7 @@ async def test_execute_query_reuse_fetches_once_then_passes_cached_schema(
         inst.__aexit__ = AsyncMock(return_value=None)
         return inst
 
-    with patch("pipefy_sdk.base_client.Client") as mock_client_cls:
+    with patch("pipefy_sdk.graphql_executor.Client") as mock_client_cls:
         first, second = _make_context_client(), _make_context_client()
         first.schema = cached_schema
         second.schema = None
@@ -145,7 +145,7 @@ async def test_execute_query_bubbles_up_execute_errors_unchanged(valid_settings)
     mock_session = AsyncMock()
     mock_session.execute = AsyncMock(side_effect=expected_error)
 
-    with patch("pipefy_sdk.base_client.Client") as mock_client_cls:
+    with patch("pipefy_sdk.graphql_executor.Client") as mock_client_cls:
         mock_client_cls.return_value.__aenter__ = AsyncMock(return_value=mock_session)
         mock_client_cls.return_value.__aexit__ = AsyncMock(return_value=None)
